Move to/from_big_compact into core endian
