+++ core/silkworm/common/endian_test.cpp
@@ -18,6 +18,8 @@
 
 #include <catch2/catch.hpp>
 
+#include <silkworm/common/util.hpp>
+
 namespace silkworm::endian {
 
 TEST_CASE("16-bit Endian") {
@@ -71,6 +73,70 @@ TEST_CASE("64-bit Endian") {
 
     uint64_t le{load_little_u64(bytes)};
     CHECK(le == 0xf0debc9a78563412);
+}
+
+static const std::string hex_endian_swap(const std::string native_hex) {
+    std::string ret{};
+    for (unsigned int i = 0; i < native_hex.length(); i += 2) {
+        ret.insert(0, native_hex.substr(i, 2));
+    }
+    return ret;
+}
+
+TEST_CASE("Block as key and compact form") {
+    const std::string block_number_hex{"000000005485ffde"};  // i.e. 1418067934
+    const std::string block_number_hex_rev{hex_endian_swap(block_number_hex)};
+
+    auto block_number{std::stoull(block_number_hex, nullptr, 16)};
+    REQUIRE(block_number == 1418067934u);
+
+    SECTION("Block number as key") {
+        // Check the sequence of bytes in memory
+        ByteView block_number_view(reinterpret_cast<uint8_t*>(&block_number), sizeof(uint64_t));
+
+#if SILKWORM_BYTE_ORDER == SILKWORM_LITTLE_ENDIAN
+        // Check we've switched to native endianness
+        CHECK(to_hex(block_number_view) == block_number_hex_rev);
+#else
+        // Check our hex form matches input form
+        CHECK(to_hex(block_number_view) == block_number_hex);
+#endif
+
+        alignas(uint64_t) uint8_t block_number_as_key[8];
+        store_big_u64(&block_number_as_key[0], block_number);
+
+        // Check data value is byte swapped if endianness requires
+        auto block_number_from_key{*reinterpret_cast<uint64_t*>(block_number_as_key)};
+
+#if SILKWORM_BYTE_ORDER == SILKWORM_LITTLE_ENDIAN
+        CHECK(block_number_from_key != block_number);
+#else
+        CHECK(block_number_from_key == block_number);
+#endif
+        CHECK(endian::be::uint64(block_number_from_key) == block_number);
+    }
+
+    SECTION("Block number as compact") {
+        // Convert block number to compact and check initial zeroes are stripped
+        auto block_number_compact_bytes{to_big_compact(block_number)};
+        CHECK(to_hex(block_number_compact_bytes) == "5485ffde");
+        // Convert back and check
+        auto block_number_from_compact{from_big_compact(block_number_compact_bytes)};
+        CHECK(block_number_from_compact == block_number);
+        // Try compact empty bytes
+        Bytes empty_bytes{};
+        CHECK(zeroless_view(empty_bytes).empty() == true);
+        // Try compact zeroed bytes
+        Bytes zeroed_bytes(2, 0);
+        CHECK(zeroless_view(zeroed_bytes).empty() == true);
+        // Compact block == 0
+        CHECK(to_big_compact(0).empty() == true);
+        // Try retrieve a compacted value from an empty Byte string
+        CHECK(from_big_compact(Bytes()) == 0u);
+        // Try retrieve a compacted value from a too large Byte string
+        Bytes extra_long_bytes(sizeof(uint64_t) + 1, 0);
+        CHECK(from_big_compact(extra_long_bytes) == std::nullopt);
+    }
 }
 
 }  // namespace silkworm::endian
