+++ node/silkworm/db/util.cpp
@@ -43,34 +43,6 @@ Bytes block_key(uint64_t block_number, const uint8_t (&hash)[kHashLength]) {
     return key;
 }
 
-Bytes to_big_compact(const uint64_t value) {
-    if (!value) {
-        return {};  // All bytes are zero
-    }
-    auto value_be_bytes{block_key(value)};
-    auto compact_view{zeroless_view(value_be_bytes)};
-    return {compact_view.data(), compact_view.length()};
-}
-
-uint64_t from_big_compact(const ByteView& data) {
-    // Important ! We can't have a string of bytes wider than an uint64_t
-    if (data.length() > sizeof(uint64_t)) {
-        throw std::invalid_argument(std::string(__FUNCTION__) + " : Data too wide");
-    }
-
-    uint64_t ret{0};
-    if (data.empty()) {
-        return ret;
-    }
-
-    uint8_t num_shifts{0};
-    for (auto i = data.rbegin(); i != data.rend(); ++i) {
-        ret |= (static_cast<uint64_t>(*i) << num_shifts);
-        num_shifts += 8;
-    }
-    return ret;
-}
-
 Bytes storage_change_key(uint64_t block_number, const evmc::address& address, uint64_t incarnation) {
     Bytes res(8 + kPlainStoragePrefixLength, '\0');
     endian::store_big_u64(&res[0], block_number);
